third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-containment-sorting-expected.txt

Tests sorting in Containment view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testSorting

Profiler was disabled.
